Move SearchableManager from cpm_data to modeladminutils
I'm going to reuse it in events app.
