+++ cpm_data/models.py
@@ -10,18 +10,9 @@ from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 from wagtail.wagtailsearch import index
 
 from modeladminutils.edit_handlers import AdminModelChooserPanel
+from modeladminutils.models import SearchableManager
 from cpm_generic.constants import COUNTRIES
 from cpm_generic.models import TranslatedField
-
-from cpm_data.queryset import SearchableQuerySet
-
-
-class BaseSearchableManager(models.Manager):
-    def get_queryset(self):
-        return SearchableQuerySet(self.model)
-
-
-SearchableManager = BaseSearchableManager.from_queryset(SearchableQuerySet)
 
 
 class Film(index.Indexed, ClusterableModel):
