+++ modeladminutils/models.py
@@ -1,1 +1,14 @@
-from __future__ import unicode_literals
+from django.db import models
+
+from modeladminutils.queryset import SearchableQuerySet
+
+
+__all__ = ['SearchableManager']
+
+
+class BaseSearchableManager(models.Manager):
+    def get_queryset(self):
+        return SearchableQuerySet(self.model)
+
+
+SearchableManager = BaseSearchableManager.from_queryset(SearchableQuerySet)
